+++ sdks/python/apache_beam/internal/pipeline/context.py
