+++ sdks/python/apache_beam/pvalue.py
@@ -57,7 +57,7 @@ if TYPE_CHECKING:
   from apache_beam.transforms.core import Windowing
   from apache_beam.pipeline import AppliedPTransform
   from apache_beam.pipeline import Pipeline
-  from apache_beam.pipeline.context import PipelineContext
+  from apache_beam.internal.pipeline.context import PipelineContext
 
 __all__ = [
     'PCollection',
