+++ sdks/python/apache_beam/transforms/environments_test.py
@@ -26,7 +26,7 @@ import logging
 import unittest
 
 from apache_beam.options.pipeline_options import PortableOptions
-from apache_beam.pipeline import context as pipeline_context
+from apache_beam.internal.pipeline import context as pipeline_context
 from apache_beam.portability import common_urns
 from apache_beam.transforms import environments
 from apache_beam.transforms.environments import DockerEnvironment
@@ -40,7 +40,7 @@ from apache_beam.transforms.environments import SubprocessSDKEnvironment
 
 class RunnerApiTest(unittest.TestCase):
   def test_environment_encoding(self):
-    from apache_beam.pipeline import context as pipeline_context
+    from apache_beam.internal.pipeline import context as pipeline_context
     for environment in (DockerEnvironment(),
                         DockerEnvironment(container_image='img'),
                         DockerEnvironment(capabilities=['x, y, z']),
@@ -75,7 +75,7 @@ class RunnerApiTest(unittest.TestCase):
     self.assertIn(common_urns.primitives.TO_STRING.urn, sdk_capabilities)
 
   def test_default_capabilities(self):
-    from apache_beam.pipeline import context as pipeline_context
+    from apache_beam.internal.pipeline import context as pipeline_context
     environment = DockerEnvironment.from_options(
         PortableOptions(sdk_location='container'))
     context = pipeline_context.PipelineContext()
